+++ packages/natives/native/index.js
@@ -24,7 +24,7 @@ export const Shell = nativeBindings.Shell;
 
 // functions
 export const __ompInstallTokioRuntime = nativeBindings.__ompInstallTokioRuntime;
-export const __piNativesV16_4_6 = nativeBindings.__piNativesV16_4_6;
+export const __piNativesV16_4_8 = nativeBindings.__piNativesV16_4_8;
 export const astEdit = nativeBindings.astEdit;
 export const astGrep = nativeBindings.astGrep;
 export const astMatch = nativeBindings.astMatch;